not collide from inside shoot pt.1
