+++ Content.Shared/SS220/Forcefield/Figures/ForcefieldEllipse.cs
@@ -1,0 +1,292 @@
+using Robust.Shared.Physics.Collision.Shapes;
+using Robust.Shared.Serialization;
+using System.Numerics;
+
+namespace Content.Shared.SS220.Forcefield.Figures;
+
+[Serializable, NetSerializable]
+[DataDefinition]
+public sealed partial class ForcefieldEllipse : IForcefieldFigure
+{
+    [DataField]
+    public float Width
+    {
+        get => _width;
+        set
+        {
+            _width = value;
+            Dirty = true;
+        }
+    }
+    private float _width = 6f;
+
+    [DataField]
+    public float Height
+    {
+        get => _height;
+        set
+        {
+            _height = value;
+            Dirty = true;
+        }
+    }
+    private float _height = 8f;
+
+    [DataField]
+    public float Thickness
+    {
+        get => _thickness;
+        set
+        {
+            _thickness = value;
+            Dirty = true;
+        }
+    }
+    private float _thickness = 0.33f;
+
+    [DataField]
+    public Angle Angle
+    {
+        get => _angle;
+        set
+        {
+            _angle = value;
+            Dirty = true;
+        }
+    }
+    private Angle _angle = default;
+
+    [DataField]
+    public Vector2 Offset
+    {
+        get => _offset;
+        set
+        {
+            _offset = value;
+            Dirty = true;
+        }
+    }
+    private Vector2 _offset = default;
+
+    [DataField]
+    public int Segments
+    {
+        get => _segments;
+        set
+        {
+            _segments = value;
+            Dirty = true;
+        }
+    }
+    private int _segments = 64;
+
+    public Angle OwnerRotation
+    {
+        get => _ownerRotation;
+        set
+        {
+            _ownerRotation = value;
+            Dirty = true;
+        }
+    }
+    private Angle _ownerRotation = default;
+    public bool Dirty { get; set; }
+    public Vector2[] InnerPoints { get; private set; } = [];
+    public Vector2[] OuterPoints { get; private set; } = [];
+
+    private readonly Ellipse _innerEllipse = new();
+    private readonly Ellipse _centralEllipse = new();
+    private readonly Ellipse _outerEllipse = new();
+
+    public ForcefieldEllipse(
+        float width,
+        float height,
+        float thickness,
+        Angle angle = default,
+        Vector2 offset = default,
+        int segments = 64
+    )
+    {
+        Width = width;
+        Height = height;
+        Thickness = thickness;
+        Angle = angle;
+        Offset = offset;
+        Segments = segments;
+
+        Refresh();
+    }
+
+    public ForcefieldEllipse()
+    {
+        Refresh();
+    }
+
+    public void Refresh()
+    {
+        RefreshEllipses();
+
+        InnerPoints = _innerEllipse.GetPoints(Segments);
+        OuterPoints = _outerEllipse.GetPoints(Segments);
+
+        Dirty = false;
+    }
+
+    private void RefreshEllipses()
+    {
+        var angle = OwnerRotation + Angle;
+
+        _centralEllipse.Width = Width;
+        _centralEllipse.Height = Height;
+        _centralEllipse.Angle = angle;
+        _centralEllipse.Offset = Offset;
+
+        var widthHeightOffset = Thickness / 2;
+
+        _innerEllipse.Width = Width - widthHeightOffset;
+        _innerEllipse.Height = Height - widthHeightOffset;
+        _innerEllipse.Angle = angle;
+        _innerEllipse.Offset = Offset;
+
+        _outerEllipse.Width = Width + widthHeightOffset;
+        _outerEllipse.Height = Height + widthHeightOffset;
+        _outerEllipse.Angle = angle;
+        _outerEllipse.Offset = Offset;
+    }
+
+    public IEnumerable<IPhysShape> GetShapes()
+    {
+        var result = new List<IPhysShape>();
+
+        for (var i = 0; i < Segments; i++)
+        {
+            var shape = new PolygonShape();
+            shape.Set(new List<Vector2>([InnerPoints[i], OuterPoints[i], OuterPoints[i + 1], InnerPoints[i + 1]]));
+
+            result.Add(shape);
+        }
+
+        return result;
+    }
+
+    public IEnumerable<Vector2> GetTrianglesVerts()
+    {
+        var verts = new List<Vector2>();
+
+        for (var i = 0; i < Segments; i++)
+        {
+            verts.Add(InnerPoints[i]);
+            verts.Add(OuterPoints[i]);
+            verts.Add(OuterPoints[i + 1]);
+
+            verts.Add(InnerPoints[i]);
+            verts.Add(InnerPoints[i + 1]);
+            verts.Add(OuterPoints[i + 1]);
+        }
+
+        return verts;
+    }
+
+    public static Vector2[] GetEllipsePoints(float width, float height, Angle angle = default, int segments = 64, bool clockwise = true)
+    {
+        if (segments <= 0)
+            throw new ArgumentException("The number of segments must be possitive.", nameof(segments));
+        if (width < 0)
+            throw new ArgumentException("The width cannot be negative.", nameof(width));
+        if (height < 0)
+            throw new ArgumentException("The height cannot be negative.", nameof(height));
+
+        var points = new List<Vector2>();
+
+        var rotationMatrix = Matrix3x2.CreateRotation((float)angle.Theta);
+        var angleStep = 2 * Math.PI / segments;
+        for (var i = 0; i <= segments; i++)
+        {
+            var pointAngle = i * angleStep;
+            if (clockwise)
+                pointAngle = -pointAngle;
+
+            var x = (float)(width / 2 * Math.Cos(pointAngle));
+            var y = (float)(height / 2 * Math.Sin(pointAngle));
+            var point = new Vector2(x, y);
+            point = Vector2.Transform(point, rotationMatrix);
+
+            points.Add(point);
+        }
+
+        return [.. points];
+    }
+
+    public bool IsInside(Vector2 point)
+    {
+        return _centralEllipse.IsInside(point);
+    }
+
+    [Serializable, NetSerializable]
+    private sealed class Ellipse()
+    {
+        public float Width
+        {
+            get => _width;
+            set
+            {
+                if (value < 0)
+                    throw new ArgumentException("The width cannot be negative.", nameof(Width));
+
+                _width = value;
+            }
+        }
+        private float _width;
+        public float Height
+        {
+            get => _height;
+            set
+            {
+                if (value < 0)
+                    throw new ArgumentException("The height cannot be negative.", nameof(Height));
+
+                _width = value;
+            }
+        }
+        private float _height;
+        public Angle Angle = default;
+        public Vector2 Offset = default;
+
+        public Vector2[] GetPoints(int segments = 64, bool clockwise = true)
+        {
+            if (segments <= 0)
+                throw new ArgumentException("The number of segments must be possitive.", nameof(segments));
+
+            var points = new List<Vector2>();
+
+            var rotationMatrix = Matrix3x2.CreateRotation((float)Angle.Theta);
+            var angleStep = 2 * Math.PI / segments;
+            for (var i = 0; i <= segments; i++)
+            {
+                var pointAngle = i * angleStep;
+                if (clockwise)
+                    pointAngle = -pointAngle;
+
+                var x = (float)(Width / 2 * Math.Cos(pointAngle));
+                var y = (float)(Height / 2 * Math.Sin(pointAngle));
+                var point = new Vector2(x, y);
+                point = Vector2.Transform(point, rotationMatrix);
+
+                points.Add(point);
+            }
+
+            return [.. points];
+        }
+
+        public bool IsInside(Vector2 point)
+        {
+            var rotationMatrix = Matrix3x2.CreateRotation((float)Angle.Opposite().Theta);
+            point = Vector2.Transform(point, rotationMatrix);
+            point -= Offset;
+
+            var normalizedX = point.X / Width / 2;
+            var normalizedY = point.Y / Width / 2;
+            return normalizedX * normalizedX + normalizedY * normalizedY <= 1;
+        }
+    }
+}
